+++ packages/sdks/snippets/react/src/routes/AnnouncementBar.tsx
