+++ packages/sdks/snippets/react/src/routes/App.tsx
